requirements.txt
openai
python-dotenv
pandas
openpyxl
Pillow
xlrd
streamlit